Refactor and unify result workflows (#684)
This commit is pure refactoring of the result workflows in dpf post.

The goal is to centralize building the workflow for the different simulation types. Before this PR there was a lot of code duplication in the different simulation types. The old code was quite hard to follow, because the creation of the sub-workflows and their connection was mixed. To simplify the code I took the following approach:

Convert all the sub-steps (averaging, principal, equivalent etc) into workflows. This simplifies the connection of the different sub-workflows. Previously a mix of operators, list of operators and (sub)workflows was used.
The creation of the overall workflow is split into two steps:
Create all the sub-workflows and put them into a ResultWorkflows object. sub-workflows that are not needed are None,
Connect the sub-workflows to get a single big workflow
Everything is moved to free functions so it is easy to reuse different parts of the workflow.
Moved component related logic from Simulation class to free functions.