Average per body (#690)
Adds an averaging_config to the _get_results and _get_results_workflow functions of

static_mechanical_simulation
transient_mechanical_simulation
modal_mechanical_simulation
harmonic_mechanical_simulation
By default the averaging happens across bodies, which is the current behavior. If average_per_body is activated, the results are averaged per body and the resulting Dataframe/FieldsContainer contains separate fields for each body (with the a tuple of the selected properties with which a body is identified). The properties by which bodies are identified are configurable. The tests use the properties "mat", "apdl_element_type", "elshape" and "apdl_real_id". The flag is currently not exposed for the results functions (stress, stress_nodal etc). The mesh support of the fields is always the full mesh.